feat: move state and expose what's needed
