+++ src/console/src/cdn/strategies_impls/storage.rs
@@ -9,7 +9,7 @@ use crate::cdn::strategies_impls::cdn::CdnHeap;
 use crate::certification::cert::update_certified_data;
 use candid::Principal;
 use ic_certification::HashTree;
-use junobuild_auth::runtime::pruned_labeled_sigs_root_hash_tree;
+use junobuild_auth::pruned_labeled_sigs_root_hash_tree;
 use junobuild_cdn::storage::errors::{
     JUNO_CDN_STORAGE_ERROR_CANNOT_GET_ASSET_UNKNOWN_REFERENCE_ID,
     JUNO_CDN_STORAGE_ERROR_CANNOT_INSERT_ASSET_ENCODING_UNKNOWN_REFERENCE_ID,
