+++ src/libs/auth/src/delegation/get.rs
@@ -1,5 +1,5 @@
 use crate::delegation::seed::calculate_seed;
-use crate::services::read_state;
+use crate::state::services::read_state;
 use crate::strategies::AuthCertificateStrategy;
 use crate::types::interface::{
     Delegation, GetDelegationArgs, GetDelegationResponse, SignedDelegation,
