+++ src/libs/auth/src/lib.rs
@@ -1,7 +1,7 @@
 mod constants;
 pub mod delegation;
-mod memory;
-pub mod runtime;
-mod services;
+mod state;
 pub mod strategies;
 pub mod types;
+
+pub use state::runtime::*;
