+++ src/libs/satellite/src/certification/strategy_impls.rs
@@ -1,7 +1,7 @@
 use crate::auth::store::get_salt;
 use crate::certification::cert::update_certified_data;
 use ic_certification::{Hash, HashTree};
-use junobuild_auth::runtime::pruned_labeled_sigs_root_hash_tree;
+use junobuild_auth::pruned_labeled_sigs_root_hash_tree;
 use junobuild_auth::strategies::AuthCertificateStrategy;
 use junobuild_auth::types::core::Salt;
 use junobuild_storage::runtime::certified_assets_root_hash;
